refactor(web): dedupe grid minute-height math, relocate visible-start helper
Extracts the clientHeight/TIMED_VISIBLE_HOURS/60 conversion (previously
duplicated across the Week and Day scroll-to-now hooks) into a single
getMinuteHeight helper, and moves getVisibleGridStartMinute next to its
sibling default-time helper in draft.util.ts instead of living inline in
EventForm.tsx.
