+++ packages/web/src/views/Week/hooks/grid/useScroll.ts
@@ -1,15 +1,17 @@
 import { type MutableRefObject, useCallback, useEffect } from "react";
-import { getCurrentMinute } from "@web/common/utils/grid/grid.util";
-import { TIMED_VISIBLE_HOURS } from "@web/grid/grid.constants";
+import {
+  getCurrentMinute,
+  getMinuteHeight,
+} from "@web/common/utils/grid/grid.util";
 import { useAppShortcut } from "@web/shortcuts/useAppShortcut";
 
 export const useScroll = (
   timedGridRef: MutableRefObject<HTMLElement | null>,
 ) => {
   const scrollToNow = useCallback(() => {
-    const gridRowHeight =
-      (timedGridRef.current?.clientHeight || 0) / TIMED_VISIBLE_HOURS;
-    const minuteHeight = gridRowHeight / 60;
+    const minuteHeight = getMinuteHeight(
+      timedGridRef.current?.clientHeight || 0,
+    );
 
     const buffer = 150;
     const top = getCurrentMinute() * minuteHeight - buffer;
